refactor: Code cleanup and optimization
Code quality improvements:
- Remove unused imports across 40 files
- Optimize Javadoc formatting (* → * <p>)
- Replace manual getters with Lombok @Getter annotations
- Remove unnecessary 'throws Exception' from test methods
- Simplify null checks and redundant code

Changes by category:

1. Main code (17 files):
   - Channel.java: Remove unused XdagBusinessHandler import
   - MessageQueue.java: Use @Getter for timerThread
   - LayeredStats.java: Use Lombok @Getter for network/application
   - ExampleEventHandler.java: Optimize Javadoc, simplify null check
   - Multiple message classes: Remove unused imports

2. Test code (23 files):
   - XdagBusinessHandlerTest.java: Remove 'throws Exception' (15 methods)
   - Remove unused imports across all test files
   - Simplify test setup code

3. New test utilities (4 files):
   - Add MockDnsResolver for in-memory DNS testing
   - Add MockableLookUpTxt for mock/real DNS switching
   - Add MockDnsTest with 9 comprehensive tests
   - Add README.md with usage guide

4. Test results (5 files):
   - Discovery test reports from recent runs

Impact:
✅ All 471 tests passing (100% pass rate)
✅ Code is cleaner and more maintainable
✅ Reduced boilerplate with Lombok
✅ Better Javadoc formatting
✅ No functional changes

Test results:
  Tests run: 471
  Failures: 0
  Errors: 0
  Skipped: 0
  Time: ~17s

+++ src/test/java/io/xdag/p2p/config/UpgradeControllerTest.java
@@ -97,7 +97,8 @@ public class UpgradeControllerTest {
         int packetId = 1;
         int packetSize = body.length;
         short packetVersion = (short)0; // XdagFrame.VERSION
-        byte compressType = (byte) (cfg.isEnableFrameCompression() ? XdagFrame.COMPRESS_SNAPPY : XdagFrame.COMPRESS_NONE);
+        byte compressType =
+            cfg.isEnableFrameCompression() ? XdagFrame.COMPRESS_SNAPPY : XdagFrame.COMPRESS_NONE;
 
         XdagFrame frame = new XdagFrame(packetVersion, compressType, packetType, packetId, packetSize, Math.min(packetSize, cfg.getNetMaxFrameBodySize()), Bytes.wrap(body, 0, Math.min(packetSize, cfg.getNetMaxFrameBodySize())).toArray());
